refactor: 결과 페이지에서 결과를 불러오는 코드 훅 분리
